fix modify sauce

+++ backend/controllers/Sauces.js
@@ -33,27 +33,35 @@ exports.createSauce = (req, res, next) => {
 };
 
 exports.modifySauce = (req, res, next) => {
-	const sauceObject = req.file
-		? {
-				...JSON.parse(req.body.sauce),
-				imageUrl: `${req.protocol}://${req.get("host")}/images/${
-					req.file.filename
-				}`,
-		  }
-		: { ...req.body };
 	Sauce.findOne({ _id: req.params.id })
 		.then((sauce) => {
-			const filename = sauce.imageUrl.split("/images/")[1];
-			fs.unlink(`images/${filename}`, () => {
+			if (req.file) {
+				const sauceObject = {
+					...JSON.parse(req.body.sauce),
+					imageUrl: `${req.protocol}://${req.get("host")}/images/${
+						req.file.filename
+					}`,
+				}
+				const filename = sauce.imageUrl.split("/images/")[1];
+				fs.unlink(`images/${filename}`, () => {
+					Sauce.updateOne(
+						{ _id: req.params.id },
+						{ ...sauceObject, _id: req.params.id }
+					)
+						.then(() =>
+							res.status(200).json({ message: "Sauce modifié !" })
+						)
+						.catch((error) => res.status(400).json({ error }));
+				});	
+			} 
 				Sauce.updateOne(
 					{ _id: req.params.id },
-					{ ...sauceObject, _id: req.params.id }
+					{ ...req.body, _id: req.params.id }
 				)
 					.then(() =>
 						res.status(200).json({ message: "Sauce modifié !" })
 					)
 					.catch((error) => res.status(400).json({ error }));
-			});
 		})
 		.catch((error) => res.status(500).json({ error }));
 };
